docs(agent-distribution): record the publisher-lane supersession (FR-1)
Rewrites the "Convergence" transport-law section: the runtime binary is
now documented as the FLOOR transport (a machine that never fetches is
still fully correct), with the publisher lane as a second, boot-only
transport that consciously supersedes 796ff1f08's binary-only conclusion
while preserving the process-lifetime immutability invariant that
conclusion protected. Documents the boot sequence, the publisher CI
pipeline, rollback-by-republish, the two-release key-rotation
choreography, /meta channel advertisement, worker telemetry (never
desired state), and agent_catalog_override's fold as the sanctioned
per-tenant lever.

Also splits new /meta agent-catalog tests into a sibling
test_meta_agent_catalog.py (kept test_meta_endpoint.py under the repo's
600-line cap) and nudges two allowlisted files' line counts for this
rung's additions.
