+++ scripts/max_lines_allowlist.txt
@@ -43,7 +43,7 @@ anyharness/sdk/src/reducer/__tests__/transcript.test.ts 1721 existing SDK oversi
 anyharness/sdk/src/reducer/transcript.ts 1515 existing SDK oversized file; legacy Subagents MCP reducer branch removed in PR10
 apps/desktop/src-tauri/src/commands/ssh_tunnel.rs 729 existing desktop native oversized file
 server/proliferate/auth/identity/store.py 695 managed sandbox GitHub grant read path extends existing identity store + SG10 review: explicit crypto-secret calls retain the original readable grant flow instead of line-neutral compression
-server/proliferate/config.py 614 managed Cloud, GitHub App, and Workflow flags extend the centralized deployment settings surface; split pending
+server/proliferate/config.py 621 managed Cloud, GitHub App, and Workflow flags extend the centralized deployment settings surface; split pending; +7: publisher-lane agent_catalog_artifact_base_url/agent_catalog_channel settings (Update Flow ADR, FR-1)
 server/proliferate/db/models/cloud/agent_gateway.py 643 C-train composition: C-2 agent_auth_delivery_ack model joins the B/D cut; agent_gateway_enrollment_key model added for per-(subject,harness) scoped virtual keys (B2), then the agent_model_snapshot re-key (B4) documented why its scope carries NO unique key and what that costs the reader (the active row is plural in principle, so reads tie-break on id); AgentApiKey gains a kind column (D1); a module split is deferred to the model-taxonomy pass rather than done mid-stack; typed-config write gate loosens ck_agent_auth_selection_api_key_shape with a why-not-tighter comment; SG12 keeps the base docstring intact while the direct wall-clock owner import adds one line
 server/proliferate/server/accounts/identity/api.py 514 desktop-surface provider callbacks now dispatch to an HTML handoff/error page instead of a raw 302 via a new _provider_callback_response helper (PRO-111); split pending
 server/proliferate/server/billing/stripe_webhooks.py 665 W-F2 period-boundary gate adds one guard clause (a billing_reason check plus its drop report and the hold-clear that must still run) to _handle_invoice_paid; the reporter and its context assembly live in webhook_drops.py. W-F5 drop visibility: the intake path was already 595/600 before this change, so threading the Stripe event id through the five handlers and adding one report call per silent early return cannot fit under the cap. The drop vocabulary, levels, paid-detection, and all per-reason context assembly were split out to webhook_drops.py to keep the growth to call sites only; splitting the handlers themselves is deferred to the billing service taxonomy pass rather than done inside an observability-only fix. Adversarial-review fix pass (MAJOR 1) adds a two-line reporter call at the dispatch fall-through's 13th silent drop (unhandled_event_type); billing error normalization changes only the direct error import
@@ -79,7 +79,7 @@ apps/packages/product-client/src/domain/chats/transcript/transcript-presentation
 apps/packages/product-client/src/domain/workspaces/cloud-work-inventory.test.ts 859 relocated existing oversized domain test; no growth
 apps/packages/product-client/src/domain/sessions/activity.test.ts 660 relocated existing oversized domain test; no growth
 apps/packages/product-client/src/domain/sessions/intents/session-intent.test.ts 606 relocated existing oversized domain test; no growth
-anyharness/crates/anyharness-lib/src/app/mod.rs 674 AnyHarness app wiring (+gateway model planner wiring; the gateway model resolver + probe-engine store wiring is gone with the old-chain deletion, A9) (+session mutation admission purge/retention wiring, +goal/loop/activity/feed services, +two-phase workflow-run and workflow-workspace wiring, +worktree restore runtime wiring; local materialization service wiring is now the materialization runtime valve, moved to app/materialization.rs per grid PR 7; -3: WorkspaceService construction/pass-through deleted, grid PR 8; net 0 in grid PR 9: WorkspaceRetireService joins AppState while the preflight-checker/purge/retire constructions move to the app/workspaces.rs wiring family); PR #1753 extracts process/runtime configuration helpers to app/config.rs while adding Workspace MCP composition; PR #3 moves Agent Operations graph construction to app/agent_operations.rs; PR10 removes legacy Subagents MCP wiring while adding live product-context composition ; -24: R1 archiving unwires WorkspaceRetireService and WorkspaceRetentionService (fields, constructions, the retention startup pass and the WorktreeRetentionPolicyStore import) ; +12: R4 archiving wires WorkspaceArchiveService (import, field, deps pass-through, construction, and the cfg(not(test)) sweep boot pass) ; +7 net: R5 wires QuiescePlanes and the deletion-family purge construction while the clippy fix drops the dead session_delete_workflow argument ; -10: workflows gen-2 PR1 unwires the gen-1 two-phase workflow-run and workflow-workspace services ; +26: PR4 wires the WorkflowManager (construction, extension registration, boot fence pass)
+anyharness/crates/anyharness-lib/src/app/mod.rs 675 AnyHarness app wiring (+gateway model planner wiring; the gateway model resolver + probe-engine store wiring is gone with the old-chain deletion, A9) (+session mutation admission purge/retention wiring, +goal/loop/activity/feed services, +two-phase workflow-run and workflow-workspace wiring, +worktree restore runtime wiring; local materialization service wiring is now the materialization runtime valve, moved to app/materialization.rs per grid PR 7; -3: WorkspaceService construction/pass-through deleted, grid PR 8; net 0 in grid PR 9: WorkspaceRetireService joins AppState while the preflight-checker/purge/retire constructions move to the app/workspaces.rs wiring family); PR #1753 extracts process/runtime configuration helpers to app/config.rs while adding Workspace MCP composition; PR #3 moves Agent Operations graph construction to app/agent_operations.rs; PR10 removes legacy Subagents MCP wiring while adding live product-context composition ; -24: R1 archiving unwires WorkspaceRetireService and WorkspaceRetentionService (fields, constructions, the retention startup pass and the WorktreeRetentionPolicyStore import) ; +12: R4 archiving wires WorkspaceArchiveService (import, field, deps pass-through, construction, and the cfg(not(test)) sweep boot pass) ; +7 net: R5 wires QuiescePlanes and the deletion-family purge construction while the clippy fix drops the dead session_delete_workflow argument ; -10: workflows gen-2 PR1 unwires the gen-1 two-phase workflow-run and workflow-workspace services ; +26: PR4 wires the WorkflowManager (construction, extension registration, boot fence pass) ; +1: R5 (Update Flow ADR, FR-1) swaps CatalogSyncService::from_bundled() for the boot-fetch-aware from_bundled_and_staged_via_env(&runtime_home)
 anyharness/crates/anyharness-lib/src/domains/agents/catalog/service_tests.rs 605 B-2 validation-cut rewrote the refusal assertions in place (single SESSION_MODEL_UNSUPPORTED); shrinks when the legacy gateway-models routes delete (C-track cutover)
 anyharness/crates/anyharness-lib/src/domains/loops/runtime.rs 802 loops runtime — native write path + emulated Codex scheduler (+2b derived-safe classification comment at the fire seam; fire executor moved to its owning module in PR5)
 anyharness/crates/anyharness-lib/src/live/sessions/actor/startup.rs 650 session startup (+goal/loop capability parsing, roster reconcile-on-attach); R3: +1 line initializing the new pending_stop_response field on SessionActor construction; observability PR9: +2 session.established record fields ; +5: pr11b threads the interaction hooks into the event sink at construction
